isole le composant d'en-tête d'une page détail Indicateur
