+++ app.territoiresentransitions.react/src/app/pages/collectivite/Indicateurs/IndicateurPredefini.tsx
@@ -1,7 +1,7 @@
 import ScrollTopButton from 'ui/buttons/ScrollTopButton';
 import {ToolbarIconButton} from 'ui/buttons/ToolbarIconButton';
 import {useIndicateur} from './useIndicateurDefinitions';
-import {HeaderIndicateur} from './Header';
+import {HeaderIndicateur} from './detail/HeaderIndicateur';
 import {IndicateurDetail} from './detail/IndicateurDetail';
 import {IndicateurCompose} from './detail/IndicateurCompose';
 import {IndicateurSidePanelToolbar} from './IndicateurSidePanelToolbar';
